Add files via upload
This is to print the Modpack version to the console so that when people are sharing their logs, we know what version of the modpack they are using. This value will have to be manually edited each update. I'm sure we can automate the creation of this file in the future.

[main/INFO]: Loading All Of Fabric 3 v2.7.3